tools/requirements.txt
pyyaml>=6.0
structlog>=24.0
selenium>=4.15
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
google-genai>=1.0
